Verify that encoded parameters are within treshold
This also removes __slots__ since it doesn't make any difference apart
from making testing harder. Since the intention is to only create one
instance of Porridge and not many, it falls outside the intended use
of __slots__. I timed this to ensure it didn't have any impact:

$ python -m timeit -n 10000 -s 'from porridge import Porridge; p = Porridge("key:secret")' 'p.boil("password")'
10000 loops, best of 3: 1.77 msec per loop
$ # that was without __slots__, now with:
$ python -m timeit -n 10000 -s 'from porridge import Porridge; p = Porridge("key:secret")' 'p.boil("password")'
10000 loops, best of 3: 1.77 msec per loop

Not kidding, exactly the same.

Closes #1.